adds top toolbar
